etl/requirements.txt
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
classify-imports==4.2.0
distlib==0.3.6
elastic-transport==8.4.0
elasticsearch==8.5.3
filelock==3.8.2
identify==2.5.11
idna==3.4
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
platformdirs==2.6.0
pre-commit==2.20.0
pre-commit-hooks==4.4.0
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
reorder-python-imports==3.9.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
toml==0.10.2
tomli==2.0.1
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
